+++ requirements.txt
@@ -4,11 +4,10 @@
 # 
 # For Windows please install AFTER all other dependencies
 # pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
-# https://github.com/TimDettmers/bitsandbytes/issues/822
-# pip install https://github.com/jllllll/bitsandbytes-windows-webui/releases/download/wheels/bitsandbytes-0.41.1-py3-none-win_amd64.whl
+
 accelerate==0.28.0
-bitsandbytes==0.41.0
-datasets==2.14.2
+bitsandbytes==0.41.1
+datasets==2.20.0
 einops==0.6.1
 gradio==3.37.0
 peft==0.4.0
